Witgen: Handle machine calls with side effects (#1388)
Cherry-picked ef6a72fcfab8a608f54b5e25cb173e51db953b69 from #1380.

With this PR, we track whether a call to a machine led to some side
effect (e.g. added a block). In that case, the processed identity should
count has having led to some progress, even if no values were returned
to the calling machine. An example would be writing values to memory,
which does not return any values and hence does not change the state of
the caller.