Remove support to `fence.i` (#1441)
`fence.i` is the instruction to flush the instructions cache on the
processor, needed only when the executable code has been modified in
memory.

But since we don't support dynamic code generation, it is better to fail
at `fence.i` than treating it as `nop` and pretending there is nothing
wrong.