Plonky3 secure publics2 (#1507)
(Replica of plonky3-secure-publics without commit bugs.) Securely
implementing public inputs in Plonky3!

No longer requires the use of fixed columns by constraining every public
value with 3 columns-- a decrementor starting from the row index of each
exposed witness, an inverse to calculate the inverse of incrementor (0
if incrementor is 0), and a selector column constrained by a series of
is_zero constraints to be 1 at the row index and 0 everywhere else.