Speed up tests (#1556)
Recent PRs (#1538, #1534) have significantly slowed down the CI, because
they run every test twice, once with the "plain" backend and once with
the `CompositeBackend`.

This PR makes sure that at least we only do the steps before (analyzing
& optimizing PIL, fixed column computation, witness computation) only
once, by making a "prepared" pipeline once and cloning it to test
different backends.

This leads to a slight speed up in the tests.