Fix registers memory used by bootloader inputs (#1655)
Fixes https://github.com/powdr-labs/powdr/pull/1652

In continuations mode, a chunk_i's first row duplicates the last row of
the previous chunk_j (if applicable). For many instructions that's fine,
but for things like `x10 <== x10 - 1;` it's not, because the instruction
ends up being executed twice and the values are just wrong in the new
chunk. Therefore when creating the bootloader inputs for chunk_i (the
later one) we need to use the second last state of the registers memory
from chunk_j, which is done in this PR.