Parameterized expressions (#1627)
This PR 'parameterise' Expressions to be able to choose when they can be
converted into a StructExpression and when they cannot. This
modification allows us to overcome ambiguities with Structs and to adopt
the standard Structs syntax:
```
struct Point{ x: int, y: int }
let p: Point = Point{ x: 1, y: 2 };
```

This PR replace #1591 to be able to include this modifications before
merging structs.

---------

Co-authored-by: chriseth <chris@ethereum.org>