Div and mul for 32 bit numbers for baby bear (#1778)
Two operations: 
- `mul` takes two 32 bit-numbers, multiply them, and outputs two 32
bit-numbers, which are the upper and lower halves of the product
- `div` takes two 32 bit-numbers, which are the dividend and divisor,
respectively, and outputs two 32 bit-numbers, which are the quotient and
the remainder, respectively

Warning:
- Although we do provide hints to calculate a unique remainder that's
less than the divisor, this is not constrained and thus witness for
`div` is not unique and not sound.

Testing:
- I created new test vectors but should be quite comprehensive (uses all
variables).

Implementation:
- Uses the affine equation `x1 * y1 + x2 = y2 * 2**32 + y3`.
- `x2` is constrained to 0 for `mul`.
- `y2` is constrained to 0 for `div`.
- This approach has the disadvantage of using five sets of columns, i.e.
`x1`, `y1`, `x2`, `y2`, and `y3`, instead of just four sets. However,
five sets of columns has the advantage of using roughly half of the
number of constraints compared to if we were to have four sets of
columns. This is because `mul` and `div` can be both defined by affine
with five sets of columns, whereas defining `mul` and `div` separately
in four sets of columns requires using the variables for different
purposes depending on the operations, and thus more constraints.
- Tl;dr: it's a trade off between one more variable (4 more columns) and
fewer constraints.