Rename trait functions (#1816)
Since we also need name and type pairs for structs, it makes sense to
rename `TraitFunction` to something more general (`NamedType` in this
case).

There are also `NamedExpressions` and everything could be combined into
a `NamedElement` or something similar that covers both cases. Personally
I prefer this option, but the discussion is open.