PIL -> rust compiler (#1416)
This is a (yet unused) component that takes an `Analyzed` and a list of
names and turns these names into `fn(u64) -> u64` to be called directly.

Not all syntax is supported yet (and no built-in functions).

The first goal is not be able to compile all constant functions.

The benchmarks show that `sqrt` is roughly 10 times faster than the
evaluator.

---------

Co-authored-by: Gastón Zanitti <gzanitti@gmail.com>
Co-authored-by: Georg Wiese <georgwiese@gmail.com>