Functions as expressions (#1837)
Adds support for symbols of function type being assigned from an
expression that is not a lambda expression.

This case is a bit more complicated because those symbols are stored as
lazy statics and not as direct `fn <name>` functions. Because of that,
any reference to it needs to be dereferenced. At the same time,
references to functions cannot be dereferenced.

We need both flavours since only `fn`-functions can be used for
recursive calls in rust.

---------

Co-authored-by: Gastón Zanitti <gzanitti@gmail.com>
Co-authored-by: Georg Wiese <georgwiese@gmail.com>