JIT support for match expressions. (#1838)
This PR supports match expression, while most of the work regards
patterns.

Unfortunately, we cannot use rust patterns directly: The types are too
incompatible. For example a `match ("abc", 7) { ("abc", 7) => ...}`
cannot be directly translated, since the rust types here are `String`
and `ibig::IBig` and they cannot be used with these literals like that,
at least not directly and in all circumstances.

The way it is implemented here is that each pattern is compiled to code
that is supposed to evaluate to an Option. If the pattern matches, the
Option is a Some-value that contains the values that are assigned to the
variables in the pattern. The function also returns a string containing
the variable names. The function calls itself recursively on recursive
data structures.

---------

Co-authored-by: Gastón Zanitti <gzanitti@gmail.com>
Co-authored-by: Georg Wiese <georgwiese@gmail.com>