JIT: proper function types. (#1866)
Rust's function types are more complicated than powdr's: It has `fn`,
`Fn` and closures (we ignore generic functions and trait functions for
now). This PR adds an
[enum](https://github.com/powdr-labs/powdr/pull/1866/files#diff-0b9128f385080144659241b59a28702f0a85a684895f59bf06f5a0b4440070fdR90)
that can hold either an `fn` or an `Fn` (which is used for closures).
This enum will be used as the rust type for a powdr function.

Rust closures that capture variables have to be handled specially: We
need to make sure that we can move the captured values into the closure
(the other option would be to use references, but we cannot say anything
about the lifetime of the captured value, so this is not viable). We do
this by first cloning the captured value into a new local variable (of
the same name), which we can then move into the closure. In order to do
this, we need to check which local variables references reach out of the
lambda and which ones stay inside the lambda. This is done by keeping
track of the current "variable height" during expression conversion.