Support any update constraint. (#1893)
The `linearize` function reduces the degree of update constraints to
linear. This is not strictly necessary and could be done at a later
stage that is not specific to asm update constraints.

This function paniced for anything that is not a simple algebraic
expression. This PR extends the functionality to any expression,
linearizing all operands of a binary operator.

Note that this will still not result in a proper expression for anything
that is more complex than a simple algebraic expression.

At some point, we should move this processing to a stage like the
condenser or even later (close to the optimizer, or even make it
backend-specific).