change toolchain to nightly (#1908)
related to [this PR](https://github.com/powdr-labs/powdr/pull/1898)
we need to change to nightly toolchain to integrate stwo

I kept the toolchain related to riscv to be "nightly-2024-08-01" as it
is handled separately in workflow, so I made the least change to make
stwo integrate-able for now.

fix some clippy issues about the comment format on some files.

---------

Co-authored-by: chriseth <chris@ethereum.org>