Add PhantomLookup and PhantomPermutation to Constr (#1934)
A miniature version of #1378, adding only annotations for a generic
lookup & permutation. This should be enough to get us to a sound version
of VADCOP.

I added the annotation to the `std::protocols::lookup::lookup` function.
We would need it to other functions in the same module as well, but I
think we can get this example working end-to-end first (in case we
decide that the annotation type needs to change for some reason).