Fix small continuations issue (#1949)
I ran into this unwrapping a None when using continuations with a very
low max degree. It's kinda hard to trigger without triggering a bunch of
other issues, but I think this fix is enough for now: if a register is
never used it is not contained in the snapshot, so it should be 0.