Rust constr enum (#1941)
`Identity` is currently a struct which acts as an union: all kinds of
identities are represented using the same struct fields, with runtime
checks that the right fields are used.

In the context of #1934 where we add new kinds of identities, it seems
advantageous to move to an enum on the Rust side as well.

Todo:
- [x] reimplement ids
- [x] minimize code duplication between lookups and permutations, if
possible
- [x] remove set_id
- [x] find better way to compare identities
- [x] fix connect