Properly re-print trait impls. (#1956)
Before this PR, the list of trait implementations was not present in
Analyzed, but we need them for re-printing the PIL (and other things).

This PR also contains a simplification in that it only keeps a list of
all trait impls, not sorted into which trait they implement. This
information is only relevant during trait impl solving, where we
re-build that data structure. Keeping the pre-sorted layout would make
re-printing in source order much more difficult.

And now it also formats paths to the current namespace properly.