Make phantom lookups work (except for multiplicity witgen) (#1953)
This PR contains some fixes to #1945. As a result, Phantom lookups /
permutations are treated exactly like native lookups / permutations in
witgen. Witness generation for the multiplicity column still works for
fixed lookups (as on `main`).

I manually searched for all uses of `Identity::Lookup` and
`Identity::Permutation` and added the phantom variants when applicable.

I also consolidated the tests a bit: The once we had were quite esoteric
(LHS all witness columns, with some hints that make the constraint work)
and didn't actually succeed if a native lookup was used. Also, we had a
bunch of `_ext` variants, but the selection of extension field or not is
done by `lookup()`, so the tests were essentially the same. Now the
tests are as follows:
- `test_data/std/lookup_via_challenges.asm` is changed to a more
realistic example of a fixed lookup. It is very similar to the old
`test_data/std/multiplicities.asm`, which I deleted.
- `test_data/std/lookup_via_challenges_range_constraint.asm` is somewhat
similar to the deleted
`test_data/std/lookup_via_challenges_ext_simple.asm` (the LHS and RHS
lists have a length of 1) and also tests that witness generation for
range constraint multiplicities works as expected (which currently
doesn't, so the test is marked as `should_panic`).