Put publics inside namespace. (#2050)
We inherited from polygon the speciality that publics are always global
even if declared inside a namespace. This PR changes that and treats
publics just as any other definition. It also does some clean-up in that
the degrees of the namespaces are first determined in the Display
function and not only while printing. This allows declarations without
degree to be used as a first declaration in a namespace.