Generalize multiplicity witgen (#2043)
(mostly cherry-picked from #1958)

This PR generalizes the multiplicity witness generation we had in
`FixedLookup` in the following ways:
- Instead of searching for a column of a specific name, takes the
witness column from the phantom lookups
- As a result, it is also able to handle arbitrarily many lookups
- The witness generation should work for all machine types which can be
connected via lookups: Fixed lookup, block machines, and secondary VMs.

A common use-case that is not yet covered is global range constraints,
as they are not represented as machine calls in witness generation. This
should be fixed in a follow-up PR.

---------

Co-authored-by: chriseth <chris@ethereum.org>